Work on setting QC output variables, update styles
