+++ include/fast5_module.h
@@ -1,8 +1,8 @@
 #ifndef FAST5_MODULE_H_
 #define FAST5_MODULE_H_
 
-#include "InputStructure.h"
-#include "OutputStructures.h"
+#include "input_parameters.h"
+#include "output_data.h"
 
 
 int generateQCForFAST5(Input_Para &_input_data, Output_FAST5 &output_data);
