+++ include/fasta_module.h
@@ -1,8 +1,8 @@
 #ifndef FASTA_MODULE_H_
 #define FASTA_MODULE_H_
 
-#include "InputStructure.h"
-#include "OutputStructures.h"
+#include "input_parameters.h"
+#include "output_data.h"
 
 int qc_fasta_files(Input_Para &_input_data, Output_FA &py_output_fa);
