+++ include/fastq_module.h
@@ -1,8 +1,8 @@
 #ifndef FASTQ_MODULE_H_
 #define FASTQ_MODULE_H_
 
-#include "InputStructure.h"
-#include "OutputStructures.h"
+#include "input_parameters.h"
+#include "output_data.h"
 
 int qc_fastq_files(Input_Para &_input_data, Output_FQ &py_output_fq);
