+++ setup.py
@@ -15,23 +15,7 @@ src_files = []
 project_dir = 'src/'
 project_src_files = []
 project_src_files.extend(glob.glob(project_dir + '*.cpp'))
-
 project_headers = glob.glob('include/*.h')
-print("Project dir: ")
-print(project_dir)
-
-print("working dir: ")
-print(os.getcwd())
-
-# Print project source files
-print("Project source files:")
-for src_file in project_src_files:
-    print("\t" + src_file)
-
-# Print project headers
-print("Project headers:")
-for header in project_headers:
-    print("\t" + header)
 
 # Set up the extension
 include_dirs = ['include']
